refactor(schemas): reorganize agent schemas into domain folders
- Split agent schemas into agent/, agent-runtime/, agent-trace/
- Atomize AgentSchema.ts into single-responsibility files
  (AgentConstraints, AgentIdSchema, AgentTextSchema, AgentCapabilitySchema, AgentPatchSchema)
- Update all cross-folder import paths
