+++ packages/schemas/src/agent-runtime/AgentRuntimeConfigSchema.ts
@@ -1,6 +1,6 @@
 import { z } from "zod/v4";
 import { AgentRuntimeSchema } from "./AgentRuntimeSchema";
-import { RuntimeConnectionSchema } from "./RuntimeConnectionSchema";
+import { RuntimeConnectionSchema } from "../RuntimeConnectionSchema";
 
 export const AgentRuntimeConfigSchema = z.object({
   id: z.string().min(1),
