+++ packages/schemas/src/agent-trace/AgentSpanSchema.ts
