+++ packages/schemas/src/agent-trace/SpanStatusSchema.ts
