+++ packages/schemas/src/agent-trace/SpanTypeSchema.ts
