+++ packages/schemas/src/agent-trace/index.ts
@@ -1,0 +1,4 @@
+export * from "./AgentRawExportSchema";
+export * from "./AgentSpanSchema";
+export * from "./SpanStatusSchema";
+export * from "./SpanTypeSchema";
