+++ packages/schemas/src/agent/AgentModeSchema.ts
