+++ packages/schemas/src/index.ts
@@ -1,10 +1,6 @@
-export * from "./AgentSchema";
-export * from "./AgentModeSchema";
-export * from "./AgentRawExportSchema";
-export * from "./AgentRunStatusSchema";
-export * from "./AgentRuntimeConfigSchema";
-export * from "./AgentRuntimeSchema";
-export * from "./AgentSpanSchema";
+export * from "./agent";
+export * from "./agent-runtime";
+export * from "./agent-trace";
 export * from "./ChecklistResultSchema";
 export * from "./DistillationRunSchema";
 export * from "./DistillationSchema";
@@ -28,8 +24,6 @@ export * from "./RuntimeConnectionSchema";
 export * from "./ScriptLanguageSchema";
 export * from "./SettingsSchema";
 export * from "./SkillSchema";
-export * from "./SpanStatusSchema";
-export * from "./SpanTypeSchema";
 export * from "./SshConnectionSchema";
 export * from "./TemplateContentTypeSchema";
 export * from "./OperationOutputItemTemplateSchema";
