+++ packages/schemas/src/pipeline/node/BuiltinNodeTypeSchema.ts
@@ -1,14 +1,12 @@
 import { z } from "zod/v4";
+import { OBJECT_NODE_TYPE_ENUM } from "./ObjectNodeTypeSchema";
+import { OPERATION_NODE_TYPE_ENUM } from "./OperationNodeTypeSchema";
+import { OUTPUT_NODE_TYPE_ENUM } from "./OutputNodeTypeSchema";
 
 export const BUILTIN_NODE_TYPE_ENUM = {
-  COMPOUND: "compound",
-  FILE: "file",
-  FOLDER: "folder",
-  GITHUB_PROJECT: "github-project",
-  OPERATION: "operation",
-  OUTPUT_PROJECT_PATH: "output-project-path",
-  OUTPUT_LOCAL_PATH: "output-local-path",
-  PROMPT: "prompt",
+  ...OBJECT_NODE_TYPE_ENUM,
+  ...OPERATION_NODE_TYPE_ENUM,
+  ...OUTPUT_NODE_TYPE_ENUM,
 } as const;
 export const BuiltinNodeTypeSchema = z.enum(BUILTIN_NODE_TYPE_ENUM);
 export type BuiltinNodeType = z.infer<typeof BuiltinNodeTypeSchema>;
