+++ packages/schemas/src/pipeline/node/index.ts
@@ -1,3 +1,6 @@
 export * from "./PipelineNodeSchema";
+export * from "./ObjectNodeTypeSchema";
+export * from "./OperationNodeTypeSchema";
+export * from "./OutputNodeTypeSchema";
 export * from "./BuiltinNodeTypeSchema";
 export * from "./MetaNodeTypeSchema";
